% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLEASE.R
\name{analyze_data}
\alias{analyze_data}
\title{Data Analysis}
\usage{
analyze_data(object)
}
\arguments{
\item{object}{An object of class "lomza".}
}
\value{
The results of the chi-squared test, t-test, and Pearson's correlation coefficient
calculation for the "lomza" object.
}
\description{
This function performs a chi-squared test, t-test, and calculates the Pearson's correlation
coefficient for all relevant numeric columns in the "lomza" object.
}
