% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLEASE.R
\name{chi_sq}
\alias{chi_sq}
\title{Chi-Squared Test}
\usage{
chi_sq(object)
}
\arguments{
\item{object}{An object of class "lomza".}
}
\value{
The chi-squared test results for each numeric column in the "lomza" object.
}
\description{
This function performs a chi-squared test on all numeric columns of the "lomza" object.
It will give a warning if there are any NA values in the column.
}
