% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLEASE.R
\name{fill_non_numeric}
\alias{fill_non_numeric}
\title{Fill Missing Values for Non-numeric Columns}
\usage{
fill_non_numeric(object, method)
}
\arguments{
\item{object}{An object of class "lomza".}
}
\value{
The "lomza" object with missing values in non-numeric columns filled.
}
\description{
This function provides multiple methods to handle missing values in non-numeric columns of the
"lomza" object. The methods include deleting the row or column containing the missing value,
filling the missing value with the most frequent value, or filling with a specified constant value.
}
