% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLEASE.R
\name{fill_numeric}
\alias{fill_numeric}
\title{Fill Missing Values for Numeric Columns}
\usage{
fill_numeric(object, method)
}
\arguments{
\item{object}{An object of class "lomza".}
}
\value{
The "lomza" object with missing values in numeric columns filled.
}
\description{
This function provides multiple methods to handle missing values in numeric columns of the
"lomza" object. The methods include deleting the row or column containing the missing value,
filling the missing value with mean, median, or mode of the column, or filling with a
specified constant value.
}
