% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLEASE.R
\name{pearson_correlation}
\alias{pearson_correlation}
\title{Pearson's Correlation Coefficient}
\usage{
pearson_correlation(object)
}
\arguments{
\item{object}{An object of class "lomza".}
}
\value{
The correlation coefficient between every pair of numeric columns in the "lomza" object.
}
\description{
This function calculates the Pearson's correlation coefficient between every pair of numeric
columns in the "lomza" object. It will give a warning if there are any NA values in the columns.
}
