% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLEASE.R
\name{summaryNEW}
\alias{summaryNEW}
\title{Extended Summary}
\usage{
summaryNEW(object)
}
\arguments{
\item{object}{An object of class "lomza".}
}
\value{
A detailed summary of the data in "lomza" object.
}
\description{
This function provides a more detailed summary of the data than the base summary function.
It only applies to numeric columns and gives warning if there are any NA values in the column.
}
