% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLEASE.R
\name{t_test}
\alias{t_test}
\title{T-Test}
\usage{
t_test(object)
}
\arguments{
\item{object}{An object of class "lomza".}
}
\value{
The t-test results for each numeric column in the "lomza" object.
}
\description{
This function performs a t-test on all numeric columns of the "lomza" object.
It will give a warning if there are any NA values in the column.
}
